rename cache file
