+++ networkapi/util/appcache.py
@@ -1,7 +1,7 @@
 import hashlib
 import logging
 
-from django.core.cache import cache as memcache
+from django.core.cache import cache as djangocache
 
 from networkapi.distributedlock import distributedlock
 
@@ -18,7 +18,7 @@ def get_cached_search(prefix, search):
     try:
         search_md5 = hashlib.md5(str(search)).hexdigest()
         key = prefix+search_md5
-        data = memcache.get(key)
+        data = djangocache.get(key)
         if data:
             log.debug("Got cached data for key %s" % key)
         return data
@@ -33,9 +33,9 @@ def set_cache_search_with_list(prefix, search, data, timeout=DEFAULT_CACHE_TIMEO
         try:
             search_md5 = hashlib.md5(str(search)).hexdigest()
             key = prefix+search_md5
-            memcache.set(key, data, timeout)
+            djangocache.set(key, data, timeout)
 
-            cached_search_md5_list = memcache.get(prefix)
+            cached_search_md5_list = djangocache.get(prefix)
             if not cached_search_md5_list:
                 cached_search_md5_list = []
 
@@ -44,7 +44,7 @@ def set_cache_search_with_list(prefix, search, data, timeout=DEFAULT_CACHE_TIMEO
 
             log.debug("Caching key %s in list %s ..." % (key, prefix))
             key = prefix
-            memcache.set(key, cached_search_md5_list, timeout)
+            djangocache.set(key, cached_search_md5_list, timeout)
         except Exception as e:
             log.error(e)
 
@@ -53,14 +53,14 @@ def delete_cached_searches_list(prefix):
 
     with distributedlock(prefix):
         try:
-            cached_search_md5_list = memcache.get(prefix)
+            cached_search_md5_list = djangocache.get(prefix)
             if cached_search_md5_list:
                 for cached_search_md5 in cached_search_md5_list:
                     key = str(prefix)+str(cached_search_md5)
                     log.debug("Deleting cache entry %s ... " % key)
-                    memcache.delete(key)
+                    djangocache.delete(key)
                 log.debug("Deleting cache list entry %s ..." % prefix)
-                memcache.delete(prefix)
+                djangocache.delete(prefix)
         except Exception as e:
             log.error(e)
             raise e
